requirements.txt
opencv-python==4.10.0.82
onvif-zeep==0.2.12
numpy==1.26
python-dotenv==1.1.1
django-cors-headers==4.7.0
celery==5.5.3
redis==6.2.0
django-celery-beat==2.8.1